async [nfc]: Move `tryUntilSuccessful` to `fetchActions` and rename.
As Greg says [1], it's always been closely tied to
`doInitialFetch`'s specific needs, with its handling of 4xx errors.
We're about to make it more so, by adding timeout logic.

[1]: https://github.com/zulip/zulip-mobile/issues/4165#issue-640704691
